+++ src/js/utils/math.util.js
@@ -14,6 +14,7 @@ const ROUNDING_MODES = Object.freeze({
 
 const ONE = 1000000
 const DECIMAL_PLACES = 6
+const MAX_ALLOWED_PERCENT = 100
 
 export class MathUtil {
   static multiply (a, b, ROUND_TYPE = ROUNDING_MODES.ROUND_UP) {
@@ -75,6 +76,20 @@ export class MathUtil {
     const result = one.plus(two)
     return result
       .toFixed(DECIMAL_PLACES)
+  }
+
+  /**
+   * Example: value = 200, percent = 15. Return - 30
+   *
+   * @param {Number|String} value
+   * @param {Number|String} percent
+   * @param {ROUND_TYPE} [ROUND_TYPE]
+   * @returns percentage of the passed value
+   */
+  static percentOfValue (value, percent, ROUND_TYPE = ROUNDING_MODES.ROUND_UP) {
+    const number = new BigNumber(value)
+    const percentValue = MathUtil.divide(percent, MAX_ALLOWED_PERCENT)
+    return MathUtil.multiply(number, percentValue)
   }
 
   static _isValidParams (op, a, b, c = 0) {