+++ src/vue/modules/dividend-form/store/index.js
@@ -2,7 +2,7 @@ import { Balance } from '../wrappers/balance'
 import { FEE_TYPES } from '@tokend/js-sdk'
 
 import { types } from './types'
-import { api } from '../_api'
+import { api, loadingDataViaLoop } from '../_api'
 import { AssetRecord } from '../wrappers/asset.record'
 
 const HORIZON_VERSION_PREFIX = 'v3'
@@ -40,11 +40,8 @@ export const actions = {
   },
   async [types.LOAD_ASSETS] ({ commit, getters }) {
     const endpoint = `/${HORIZON_VERSION_PREFIX}/assets`
-    let { data: assets } = await api().getWithSignature(endpoint, {
-      page: {
-        limit: 100,
-      },
-    })
+    let response = await api().getWithSignature(endpoint)
+    let assets = await loadingDataViaLoop(response)
 
     commit(
       types.SET_ASSETS,
@@ -104,8 +101,8 @@ export const getters = {
   [types.assets]: state => state.assets.filter(
     item => item.balance.id && item.isTransferable
   ),
-  [types.ownedAssets]: state => state.assets.filter(
-    item => item.owner === state.accountId
+  [types.ownedAssets]: state => state.assets.filter(item =>
+    item.owner === state.accountId && item.isShareSubtype
   ),
   [types.balanceHolders]: state => state.balanceHolders
     .map(item => new Balance(item)),